+++ be/Kbuild.sub
@@ -2,6 +2,7 @@ m0tr_objects += \
                   be/addb2_xc.o \
                   be/alloc_xc.o \
                   be/alloc_internal_xc.o \
+                  be/btree.o \
                   be/domain_xc.o \
                   be/extmap.o \
                   be/extmap_xc.o \
