+++ be/extmap.h
@@ -109,7 +109,7 @@
 #include "lib/types.h"     /* struct m0_uint128 */
 #include "lib/types_xc.h"  /* m0_uint128_xc */
 #include "be/tx.h"
-#include "btree/btree.h"   /* m0_btree_cursor */
+#include "be/btree.h"   /* m0_btree_cursor */
 #include "be/op.h"         /* struct m0_be_op */
 
 #include "be/extmap_internal.h"
