+++ cob/cob.h
@@ -33,7 +33,7 @@
 #include "lib/bitstring_xc.h"
 #include "fid/fid.h"
 #include "mdservice/md_fid.h"
-#include "btree/btree.h"
+#include "be/btree.h"
 
 /* import */
 struct m0_be_tx;
